+++ src/GameSceneGraph/CHANGES.md
@@ -1,5 +1,10 @@
 # Changes
 
+## Version 1.2.0
+*2021-07-24*
+
+- Adapted to version 0.10 of AST-Utilities.
+
 ## Version 1.1.0
 *2021-07-11*